partially implemented search
